pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "applied-planning"
version = "0.0.1"
description = "Applied Planning & Optimization - uFactory Lite 6 sim/control stack"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.23",
  "gymnasium>=0.29",
  "xarm-python-sdk==1.17.0",
  "jupyter>=1.0.0",
  "jupyterlab>=4.0.0",
  "matplotlib>=3.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
  "ruff>=0.5.0",
  "mypy>=1.6.0",
]
notebook = [
  "jupyter>=1.0.0",
  "jupyterlab>=4.0.0",
  "ipywidgets>=8.0.0",
  "pandas>=2.0.0",
  "pygame",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
